backend/telegram-bot/requirements.txt
pyjwt>=2.8.0
psycopg2-binary>=2.9.9
requests>=2.31.0
bcrypt>=4.1.0